# backend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "email": "gcsignin@cds-snc.ca",
      "name": "GC Sign In Team",
      "url": "https://github.com/cds-snc/gc-signin-user-self-service-webapp"
    },
    "description": "\nThis API backend service primarily interacts with [IBM Verify service](https://docs.verify.ibm.com/verify/reference).\nAPI endpoints are consumed by our frontend application to perform user authentication and profile management.\nThe API endpoints enable our custom frontend application to implement complex workflows and user interfaces without limitations.\n\n### Features\n\n* Password-based authentication\n* SMS, Email, Voice callback one-time passcode (OTP)\n",
    "title": "GC Sign In Backend API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/auth/callback": {
      "get": {
        "operationId": "callback_route_v1_auth_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lang"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Callback from IBM Verify after user authentication",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/v1/auth/legacy/login": {
      "get": {
        "description": "Initiates the OIDC authentication request to the legacy IDP for linking legacy pai",
        "operationId": "handle_legacy_login_v1_auth_legacy_login_get",
        "parameters": [
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "en",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Starts the linking flow",
        "tags": [
          "Legacy Auth"
        ]
      }
    },
...
    "/v1/auth/legacy/skip": {
      "get": {
        "description": "Handles skip flow, updating IBM Profile and redirecting to RP",
        "operationId": "handle_skip_account_linking_v1_auth_legacy_skip_get",
        "parameters": [
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lang"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Skip Linking",
        "tags": [
          "Legacy Auth"
        ]
      }
    },
    "/v1/auth/login": {
      "get": {
        "operationId": "redirect_url_v1_auth_login_get",
        "parameters": [
          {
            "in": "query",
            "name": "clientId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Clientid"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lang"
            }
          },
          {
            "in": "query",
            "name": "customparameters",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Customparameters"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Authenticate user via IBM Verify",
        "tags": [
          "Auth"
        ]
      }
    },
...
  }
}